+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,3 +1,11 @@
+# v.0.17.0 (2022-03-20)
+
+## Breaking changes
+
+- upgraded to "@cosmos-client/core": "0.44.4"
+- xchain-cosmos and xchain-thorchain now extend BaseXChainClient
+- Remove `minheight` and `maxheight` params from `CosmosSDKClient.searchTx` (params were removed from the API)
+
 # v.0.15.1 (2022-01-10)
 
 - upgrade axios dependency to latest